Ni puta idea

+++ Utilities.cpp
@@ -18,7 +18,7 @@ void swap(void* a, void* b, uint8_t elementSize) {
   uint8_t* pointerA = (uint8_t*) a;
   uint8_t* pointerB = (uint8_t*) b;
   
-  for (uint8_t i = 0; i < elementSize; i++) {
+  for (uint32_t i = 0; i < elementSize; i++) {
     aux = pointerA[i];
     
     pointerA[i] = pointerB[i];
